Cleanup, rename to opcode_tracer
